refactor: @api/

Took 14 minutes
