+++ src/app/api/prisma/add_user/route.ts
@@ -1,20 +1,10 @@
 import prisma from '@/lib/prisma'
-
-export type userType = {
-  name: string
-  email: string
-  profileImage: string
-  userName: string
-}
+import {userType} from "@api/prisma/upsert_user/route";
 
 export async function POST(req: Request) {
   const user: userType = await req.json()
-  const created_user = await prisma.user.upsert({
-    where: {
-      email:user.email,
-    },
-    create: user,
-    update: user,
+  const created_user = await prisma.user.create({
+    data: user,
   })
   return new Response(JSON.stringify({ status: 200 }))
 }
